+++ examples/graphql/java/src/test/java/io/rsocket/pythontest/ServerWithGraphQLTests.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package io.rsocket.pythontest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class ServerWithGraphQLTests {
 
 	@Test
 	void contextLoads() {